conflate Selector and Planner into the latter.
