+++ dial/planner.go
@@ -1,6 +1,13 @@
 package dial
 
-import "errors"
+import (
+	"errors"
+
+	"github.com/libp2p/go-libp2p-transport"
+)
+
+// ErrNoSuccessfulDials is returned by Select() when all dials failed.
+var ErrNoSuccessfulDials = errors.New("no successful dials")
 
 type singleBurstPlanner struct{}
 
@@ -24,4 +31,19 @@ func (*singleBurstPlanner) Next(req *Request, dialled dialJobs, last *Job, dialC
 	close(dialCh)
 
 	return nil
+}
+
+func (*singleBurstPlanner) Select(successful dialJobs) (tpt.Conn, error) {
+	if len(successful) == 0 {
+		return nil, ErrNoSuccessfulDials
+	}
+
+	// return the first successful dial.
+	for _, j := range successful {
+		if j.err == nil && j.tconn != nil {
+			return j.tconn, nil
+		}
+	}
+
+	return nil, ErrNoSuccessfulDials
 }
